requirements-dev.txt
-rrequirements.txt
coverage
django-dirtyfields
django-formtools
flake8
importanize
mock
pytest
pytest-cov
pytest-django
python-magic
sphinx
sphinx-autobuild
sphinx-rtd-theme
tox
